task: abstract slave core functions to arch layer
Abstracts slave core functions to arch layer. It doesn't
make sense to compile them, if they won't be used.
